requirements.txt
a2s
python-valve
aiohttp==3.7.4.post0
altgraph==0.17
async-timeout==3.0.1
asyncio==3.4.3
asyncio-dgram==2.1.0
attrs==20.3.0
auto-py-to-exe==2.11.0
backports.entry-points-selectable==1.1.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==7.1.2
colorama==0.4.3
cryptography==3.4.7
DateTime==4.3
discord.py==1.6.0
distlib==0.3.2
djitellopy==2.4.0
dnspython==2.1.0
docopt==0.6.2
docutils==0.18.1
EasyProcess==0.3
Eel==0.12.4
entrypoint2==0.2.3
enums.py==0.5.1
filelock==3.0.12
future==0.18.2
gd.py==0.11.0
gevent==21.8.0
gevent-websocket==0.10.1
googlesearch-python
greenlet==1.1.2
html5lib==1.1
idna==2.10
ipinfo
mcstatus==6.5.0
monotonic==1.6
multidict==5.1.0
numpy==1.21.2
opencv-python==4.5.3.56
patool==1.12
pefile==2019.4.18
platformdirs==2.2.0
playsound==1.3.0
pycparser==2.20
pydub==0.25.1
pyfiglet==0.8.post1
Pygments==2.10.0
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.2
PyNaCl==1.4.0
PySimpleGUI==4.16.0
python-dotenv==0.16.0
pytz==2021.3
pyunpack==0.2.2
requests==2.25.1
six==1.15.0
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.7.0
webencodings==0.5.1
whichcraft==0.6.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.4.0
